code review improvements
