+++ app.json
@@ -96,7 +96,7 @@
       "required": false
     },
     "MAX_S3_GET_ITERATIONS": {
-      "description": null,
+      "description": "Max retry attempts to get an S3 object",
       "required": false
     },
     "MEDIA_ROOT": {
