+++ main/s3_utils.py
@@ -1,0 +1,38 @@
+""" S3 utility functions """
+import boto3
+from django.conf import settings
+
+
+def get_s3_resource():
+    """
+    Provides an S3 resource
+
+    Returns:
+        s3.ServiceResource: An S3 resource
+    """
+    return boto3.resource(
+        "s3",
+        aws_access_key_id=settings.AWS_ACCESS_KEY_ID,
+        aws_secret_access_key=settings.AWS_SECRET_ACCESS_KEY,
+    )
+
+
+def get_s3_object_and_read(obj, iteration=0):
+    """
+    Attempts to read S3 data, and tries again up to MAX_S3_GET_ITERATIONS if it encounters an error.
+    This helps to prevent read timeout errors from stopping sync.
+
+    Args:
+        obj (s3.ObjectSummary): The S3 ObjectSummary we are trying to read
+        iteration (int): A number tracking how many times this function has been run
+
+    Returns:
+        bytes: The contents of a json file read from S3
+    """
+    try:
+        return obj.get()["Body"].read()
+    except Exception:  # pylint: disable=broad-except
+        if iteration < settings.MAX_S3_GET_ITERATIONS:
+            return get_s3_object_and_read(obj, iteration + 1)
+        else:
+            raise
