+++ main/utils.py
@@ -50,27 +50,6 @@ def now_in_utc():
     return datetime.datetime.now(tz=pytz.UTC)
 
 
-def get_s3_object_and_read(obj, iteration=0):
-    """
-    Attempts to read S3 data, and tries again up to MAX_S3_GET_ITERATIONS if it encounters an error.
-    This helps to prevent read timeout errors from stopping sync.
-
-    Args:
-        obj (s3.ObjectSummary): The S3 ObjectSummary we are trying to read
-        iteration (int): A number tracking how many times this function has been run
-
-    Returns:
-        bytes: The contents of a json file read from S3
-    """
-    try:
-        return obj.get()["Body"].read()
-    except Exception:  # pylint: disable=broad-except
-        if iteration < settings.MAX_S3_GET_ITERATIONS:
-            return get_s3_object_and_read(obj, iteration + 1)
-        else:
-            raise
-
-
 def chunks(iterable, *, chunk_size=20):
     """
     Yields chunks of an iterable as sub lists each of max size chunk_size.
