+++ pytest.ini
@@ -6,3 +6,5 @@ filterwarnings =
     ignore::DeprecationWarning
     ignore::PendingDeprecationWarning
     ignore:Failed to load HostKeys
+env =
+  CELERY_TASK_ALWAYS_EAGER=True
