+++ websites/tasks_test.py
@@ -12,7 +12,7 @@ from websites.tasks import import_ocw2hugo_course_paths, import_ocw2hugo_courses
 def test_import_ocw2hugo_course_paths(mocker, paths):
     """ mock_import_course should be called from task with correct kwargs """
     mock_import_course = mocker.patch("websites.tasks.import_ocw2hugo_course")
-    import_ocw2hugo_course_paths(paths, MOCK_BUCKET_NAME, TEST_OCW2HUGO_PREFIX)
+    import_ocw2hugo_course_paths.delay(paths, MOCK_BUCKET_NAME, TEST_OCW2HUGO_PREFIX)
     if not paths:
         mock_import_course.assert_not_called()
     else:
@@ -31,7 +31,7 @@ def test_import_ocw2hugo_courses(
     setup_s3(settings)
     mock_import_paths = mocker.patch("websites.tasks.import_ocw2hugo_course_paths.si")
     with pytest.raises(mocked_celery.replace_exception_class):
-        import_ocw2hugo_courses(  # pylint:disable=no-value-for-parameter
+        import_ocw2hugo_courses.delay(
             bucket=MOCK_BUCKET_NAME, prefix=TEST_OCW2HUGO_PREFIX, chunk_size=chunk_size
         )
     assert mock_import_paths.call_count == call_count
@@ -41,7 +41,7 @@ def test_import_ocw2hugo_courses_nobucket(mocker):
     """ import_ocw2hugo_course_paths should be called correct # times for given chunk size and # of paths """
     mock_import_paths = mocker.patch("websites.tasks.import_ocw2hugo_course_paths.si")
     with pytest.raises(TypeError):
-        import_ocw2hugo_courses(  # pylint:disable=no-value-for-parameter
+        import_ocw2hugo_courses.delay(  # pylint:disable=no-value-for-parameter
             bucket=None, prefix=TEST_OCW2HUGO_PREFIX, chunk_size=100
         )
     assert mock_import_paths.call_count == 0
